Fix StringUtils - Onto Unit Tests
